src/Tdi_3/requirements.txt
dash
dash-vtk
pillow
trimesh
pymongo
python-dotenv
requests
# opcionais (fila assíncrona):
rq
redis
# PSP (boleto) se for integrar real:
git+https://github.com/gerencianet/gn-api-sdk-python.git
